Introduce abstract BaseLoan implementation
